22/01/2023
Refactored some of AppLayout component to better manage throughout the app.
Added a new page to check the ARecords of a domain... together with a Controller and an API call to fetch all A Records.
Refactored the Dashboard with a new table and some small things.
Added the domain's current status and privacy protection to be displayed on the page.